process_dicom_file/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
matplotlib>=3.5.0
seaborn>=0.12.0
pydicom>=2.3.0
pillow>=9.0.0
scikit-image>=0.19.0
scipy>=1.10.0
nibabel>=5.0.0
tqdm>=4.64.0
diffdrr>=0.1.0 